feat: Change quantity from int to double for decimal weights
- ReceiptItem.quantity: int → double
- Database v7: recreate receipt_items with REAL quantity
- OCR parsing: use double.tryParse for quantity
- Edit dialogs: accept decimal quantity input
- Supports weight-based items like '2.18 kg'

+++ lib/data/models/receipt_item.dart
@@ -5,7 +5,7 @@ class ReceiptItem {
   final int receiptId;
   final String productName;
   final String? barcode;
-  final int quantity;
+  final double quantity;
   final double unitPrice;
   final double totalPrice;
 
@@ -48,7 +48,7 @@ class ReceiptItem {
     int? receiptId,
     String? productName,
     String? barcode,
-    int? quantity,
+    double? quantity,
     double? unitPrice,
     double? totalPrice,
   }) {
